pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown>=3.8",
    "pyyaml>=6.0.2",
]
